+++ src/components/account_book_settings/account_book_edit_modal.tsx
@@ -6,7 +6,7 @@ import { IoCloseOutline } from 'react-icons/io5';
 import { APIName } from '@/constants/api_connection';
 import APIHandler from '@/lib/utils/api_handler';
 import { ICompanySetting } from '@/interfaces/company_setting';
-import { IAccountBookForUser } from '@/interfaces/company';
+import { IAccountBookForUser } from '@/interfaces/account_book';
 import { Button } from '@/components/button/button';
 import { LocaleKey } from '@/constants/normal_setting';
 import SelectCountryDropdown from '@/components/user_settings/select_country_dropdown';
