+++ src/constants/account_book.ts
@@ -1,5 +1,4 @@
-import { IAccountBookForUser } from '@/interfaces/company';
-import { WORK_TAG } from '@/constants/company';
+import { WORK_TAG, IAccountBookForUser } from '@/interfaces/account_book';
 
 export const FAKE_COMPANY_AND_ROLE_LIST: IAccountBookForUser[] = [
   {
