+++ src/lib/utils/counterparty.ts
@@ -2,7 +2,7 @@ import { ICounterPartyEntity } from '@/interfaces/counterparty';
 import { Counterparty as PrismaCounterParty } from '@prisma/client';
 import { getTimestampNow } from '@/lib/utils/common';
 import { CounterpartyType } from '@/constants/counterparty';
-import { ICompanyEntity } from '@/interfaces/company';
+import { ICompanyEntity } from '@/interfaces/account_book';
 
 /**
  * Info: (20241111 - Murky)
