+++ src/pages/api/v1/status_info.ts
@@ -1,7 +1,7 @@
 import { STATUS_MESSAGE } from '@/constants/status_code';
 import { IResponseData } from '@/interfaces/response_data';
 import { IUser } from '@/interfaces/user';
-import { IAccountBook } from '@/interfaces/company';
+import { IAccountBook } from '@/interfaces/account_book';
 import { formatApiResponse } from '@/lib/utils/common';
 import { formatCompany } from '@/lib/utils/formatter/company.formatter';
 import { formatUser } from '@/lib/utils/formatter/user.formatter';
